Rename CRD api group to serverless
