+++ pkg/crds/main.go
@@ -42,10 +42,10 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/healthz"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 
-	apiv1alpha1 "github.com/cortexlabs/cortex/pkg/crds/apis/api/v1alpha1"
 	batch "github.com/cortexlabs/cortex/pkg/crds/apis/batch/v1alpha1"
-	apicontrollers "github.com/cortexlabs/cortex/pkg/crds/controllers/api"
+	serverless "github.com/cortexlabs/cortex/pkg/crds/apis/serverless/v1alpha1"
 	batchcontrollers "github.com/cortexlabs/cortex/pkg/crds/controllers/batch"
+	serverlesscontrollers "github.com/cortexlabs/cortex/pkg/crds/controllers/serverless"
 	//+kubebuilder:scaffold:imports
 )
 
@@ -59,7 +59,7 @@ func init() {
 	utilruntime.Must(istioscheme.AddToScheme(scheme))
 
 	utilruntime.Must(batch.AddToScheme(scheme))
-	utilruntime.Must(apiv1alpha1.AddToScheme(scheme))
+	utilruntime.Must(serverless.AddToScheme(scheme))
 	//+kubebuilder:scaffold:scheme
 }
 
@@ -165,7 +165,7 @@ func main() {
 		setupLog.Error(err, "unable to create controller", "controller", "BatchJob")
 		os.Exit(1)
 	}
-	if err = (&apicontrollers.RealtimeAPIReconciler{
+	if err = (&serverlesscontrollers.RealtimeAPIReconciler{
 		Client:        mgr.GetClient(),
 		ClusterConfig: clusterConfig,
 		Log:           ctrl.Log.WithName("controllers").WithName("api").WithName("RealtimeAPI"),
